requirements.txt
numpy==1.19.5
Flask==2.0.0
tensorflow==2.4.1
opencv-python==4.5.1.48
matplotlib==3.4.1
Pillow==7.0.0
Keras==2.4.3
scikit-learn==0.23.2
paho-mqtt==1.5.1